+++ static/app/views/insights/common/components/widgets/crashFreeSessionsChartWidget.tsx
