+++ static/app/views/insights/common/components/widgets/newAndResolvedIssueChartWidget.tsx
