+++ static/app/views/insights/common/components/widgets/releaseNewIssuesChartWidget.tsx
