+++ static/app/views/insights/common/components/widgets/releaseSessionCountChartWidget.tsx
