+++ static/app/views/insights/common/components/widgets/releaseSessionPercentageChartWidget.tsx
