+++ static/app/views/insights/common/components/widgets/sessionHealthRateChartWidget.tsx
