+++ static/app/views/insights/common/components/widgets/userHealthRateChartWidget.tsx
